Add files via upload
My basic outline of the frame, reels, and animations. 

Still needs line between the reels to complete the frame, longer reels, a middle horizontal line to show where the reels stop, and handle or button, a click animation to start the reels, updated animation to stop the reels so they dont just go off screen, etc.